csc181/requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
cffi==1.11.5
click==6.7
colorama==0.3.9
configparser==3.5.0
cryptography==2.3.1
dominate==2.3.4
enum34==1.1.6
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
futures==3.2.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysql-connector-python==8.0.11
MySQL-python==1.2.3
pycparser==2.19
pylint==1.9.3
PyMySQL==0.9.2
singledispatch==3.4.0.3
six==1.11.0
sql==0.4.0
SQLAlchemy==1.2.11
virtualenv==16.0.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
wtf==0.1
WTForms==2.2.1
